# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Castor Orçamento API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /entregas:
    get:
      operationId: listEntregas
      tags: [entregas]
      summary: Listar entregas
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      responses:
        "200":
          description: Lista de entregas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entrega"
    post:
      operationId: criarEntrega
      tags: [entregas]
      summary: Criar entrega
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarEntregaInput"
      responses:
        "200":
          description: Entrega criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entrega"
  /entregas/{id}/status:
    patch:
      operationId: updateStatusEntrega
      tags: [entregas]
      summary: Atualizar status da entrega
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pendente, em_rota, entregue, cancelado]
              required: [status]
      responses:
        "200":
          description: Status atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entrega"
  /entregas/{id}:
    patch:
      operationId: updateEntrega
      tags: [entregas]
      summary: Atualizar dados da entrega
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarEntregaInput"
      responses:
        "200":
          description: Entrega atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entrega"
    delete:
      operationId: deleteEntrega
      tags: [entregas]
      summary: Remover entrega
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Entrega removida
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
components:
  parameters:
    SessionToken:
      name: x-session-token
      in: header
      required: false
      schema:
        type: string
  schemas:
...
    Entrega:
      type: object
      properties:
        id:
          type: number
        orcamentoId:
          type: number
          nullable: true
        cliente:
          type: string
        whatsapp:
          type: string
          nullable: true
        endereco:
          type: string
          nullable: true
        produtos:
          type: string
          nullable: true
        status:
          type: string
          enum: [pendente, em_rota, entregue, cancelado]
        vendedor:
          type: string
          nullable: true
        observacoes:
          type: string
          nullable: true
        dataEntrega:
          type: string
          nullable: true
        criadoEm:
          type: string
        atualizadoEm:
          type: string
      required: [id, cliente, status]
    CriarEntregaInput:
      type: object
      properties:
        orcamentoId:
          type: number
        cliente:
          type: string
        whatsapp:
          type: string
        endereco:
          type: string
        produtos:
          type: string
        status:
          type: string
        vendedor:
          type: string
        observacoes:
          type: string
        dataEntrega:
          type: string
      required: [cliente]
...