# lib/api-zod/src/generated/types/entrega.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Castor Orçamento API
 * OpenAPI spec version: 0.1.0
 */
import type { EntregaStatus } from "./entregaStatus";

export interface Entrega {
  id: number;
  orcamentoId?: number | null;
  cliente: string;
  whatsapp?: string | null;
  endereco?: string | null;
  produtos?: string | null;
  status: EntregaStatus;
  vendedor?: string | null;
  observacoes?: string | null;
  dataEntrega?: string | null;
  criadoEm?: string;
  atualizadoEm?: string;
}